Rename the NotificationHandler to WebhookHnadler
