+++ Adyen.Test/MarketPayTests/MarketPayNotificationTest.cs
@@ -1,6 +1,6 @@
 ﻿using Adyen.Model.MarketPay;
 using Adyen.Model.MarketPay.Notification;
-using Adyen.Notification;
+using Adyen.Webhooks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Adyen.Test.MarketPayTest
@@ -12,8 +12,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountClosedNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-closed-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+            WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountCloseNotification accountCloseNotification = (AccountCloseNotification)notificationMessage;
             Assert.AreEqual(accountCloseNotification.EventType, "ACCOUNT_CLOSED");
             Assert.AreEqual(accountCloseNotification.PspReference, "TSTPSPR0001");
@@ -24,8 +24,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountCreatedNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-created-success.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountCreateNotification accountCreateNotificationMessage = (AccountCreateNotification)notificationMessage;
             Assert.AreEqual("ACCOUNT_CREATED", accountCreateNotificationMessage.EventType);
             Assert.AreEqual("000", accountCreateNotificationMessage.Error.ErrorCode);
@@ -55,8 +55,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountHolderCreatedNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-holder-created-success.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountHolderCreateNotification accountHolderCreateNotificationMessage = (AccountHolderCreateNotification)notificationMessage;
             Assert.AreEqual("ACCOUNT_HOLDER_CREATED", accountHolderCreateNotificationMessage.EventType);
             Assert.IsNotNull(accountHolderCreateNotificationMessage.Content);
@@ -69,8 +69,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountHolderVerificationNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-holder-verification.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountHolderVerificationNotification notification = (AccountHolderVerificationNotification)notificationMessage;
             Assert.AreEqual("ACCOUNT_HOLDER_VERIFICATION", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -81,8 +81,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountHolderStatusChangeNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-holder-status-change.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountHolderStatusChangeNotification notification = (AccountHolderStatusChangeNotification)notificationMessage;
             Assert.AreEqual("ACCOUNT_HOLDER_STATUS_CHANGE", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -95,8 +95,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountHolderPayoutFailNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-holder-payout-fail.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountHolderPayoutNotification notification = (AccountHolderPayoutNotification)notificationMessage;
             Assert.AreEqual("ACCOUNT_HOLDER_PAYOUT", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -110,9 +110,9 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountHolderUpdatedNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-holder-updated.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
+           WebhookHandler webhookHandler = new WebhookHandler();
 
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountHolderUpdateNotification notification = (AccountHolderUpdateNotification)notificationMessage;
             Assert.AreEqual("ACCOUNT_HOLDER_UPDATED", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -126,8 +126,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayBeneficiarySetupNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/beneficiary-setup.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             BeneficiarySetupNotification notification = (BeneficiarySetupNotification)notificationMessage;
             Assert.AreEqual("BENEFICIARY_SETUP", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -139,9 +139,9 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayScheduledRefundsNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/scheduled-refunds-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
+           WebhookHandler webhookHandler = new WebhookHandler();
 
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             ScheduledRefundsNotification notification = (ScheduledRefundsNotification)notificationMessage;
             Assert.AreEqual("SCHEDULED_REFUNDS", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -154,8 +154,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayCompensateNegativeBalanceNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/compensate-negative-balance-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             CompensateNegativeBalanceNotification notification = (CompensateNegativeBalanceNotification)notificationMessage;
             Assert.AreEqual("COMPENSATE_NEGATIVE_BALANCE", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -167,8 +167,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayPaymentFailureNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/payment-failure-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             PaymentFailureNotification notification = (PaymentFailureNotification)notificationMessage;
             Assert.AreEqual("PAYMENT_FAILURE", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -180,8 +180,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayReportAvailableNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/report-available-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             ReportAvailableNotification notification = (ReportAvailableNotification)notificationMessage;
             Assert.AreEqual("REPORT_AVAILABLE", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -192,8 +192,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayTransferFundsNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/transfer-funds-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             TransferFundsNotification notification = (TransferFundsNotification)notificationMessage;
             Assert.AreEqual("TRANSFER_FUNDS", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -211,8 +211,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountUpdatedNotificationTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-updated-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountUpdateNotification notification = (AccountUpdateNotification)notificationMessage;
             Assert.AreEqual("ACCOUNT_UPDATED", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -223,8 +223,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountFundsBelowThresholdTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-funds-below-thresold-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountFundsBelowThresholdNotification notification = (AccountFundsBelowThresholdNotification)notificationMessage;
             Assert.AreEqual("ACCOUNT_FUNDS_BELOW_THRESHOLD", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -237,9 +237,9 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountHolderStoreStatusChangeTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-holder-store-status-change-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
+           WebhookHandler webhookHandler = new WebhookHandler();
 
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountHolderStoreStatusChangeNotification notification = (AccountHolderStoreStatusChangeNotification)notificationMessage;
 
             Assert.AreEqual("ACCOUNT_HOLDER_STORE_STATUS_CHANGE", notification.EventType);
@@ -252,9 +252,9 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayAccountHolderUpcomingDeadlineTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/account-holder-upcoming-deadline-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
+           WebhookHandler webhookHandler = new WebhookHandler();
 
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             AccountHolderUpcomingDeadlineNotification notification = (AccountHolderUpcomingDeadlineNotification)notificationMessage;
 
             Assert.AreEqual("ACCOUNT_HOLDER_UPCOMING_DEADLINE", notification.EventType);
@@ -267,8 +267,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayDirectDebitInitiatedTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/direct-debit-initiated-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             DirectDebitInitiatedNotification notification = (DirectDebitInitiatedNotification)notificationMessage;
             Assert.AreEqual("DIRECT_DEBIT_INITIATED", notification.EventType);
             Assert.IsNotNull(notification.Content);
@@ -280,8 +280,8 @@ namespace Adyen.Test.MarketPayTest
         public void MarketPayRefundFundsTransferTest()
         {
             string json = GetFileContents("Mocks/marketpay/notification/refund-funds-transfer-test.json");
-            NotificationHandler notificationHandler = new NotificationHandler();
-            IGenericNotification notificationMessage = notificationHandler.HandleMarketpayNotificationJson(json);
+           WebhookHandler webhookHandler = new WebhookHandler();
+            IGenericNotification notificationMessage = webhookHandler.HandleMarketpayNotificationJson(json);
             RefundFundsTransferNotification notification = (RefundFundsTransferNotification)notificationMessage;
             Assert.AreEqual("REFUND_FUNDS_TRANSFER", notification.EventType);
             Assert.IsNotNull(notification.Content);
