--- Adyen/Notification/NotificationHandler.cs
@@ -1,19 +1,0 @@
-﻿using Adyen.Model.MarketPay.Notification;
-using Adyen.Model.Notification;
-using Adyen.Util;
-
-namespace Adyen.Notification
-{
-    public class NotificationHandler
-    {
-        public NotificationRequest HandleNotificationRequest(string jsonRequest)
-        {
-            return JsonOperation.Deserialize<NotificationRequest>(jsonRequest);
-        }
-
-        public IGenericNotification HandleMarketpayNotificationJson(string jsonRequest)
-        {
-            return JsonOperation.Deserialize<IGenericNotification>(jsonRequest);
-        }
-    }
-}
